Converted uint32_t to uint64_t wherever possible. Update mesh to use only one index buffer object and index array. Material groups simply hold an offset into the array and a count of the indices used by the group

+++ include/array.h
@@ -7,9 +7,9 @@ typedef struct array array_t;
 
 array_t* array_create(size_t capacity, size_t elem_size);
 void array_append(array_t *a, void *datum);
-void* array_at(array_t *a, uint32_t index);
+void* array_at(array_t *a, uint64_t index);
 void* array_back(array_t *a);
-void array_set(array_t *a, uint32_t index, void *datum);
+void array_set(array_t *a, uint64_t index, void *datum);
 const void* array_data(array_t *a);
 size_t array_size(array_t *a);
 void array_copy(array_t *dst, array_t *src);
